requirements.txt

lazy_imports
pandas
scikit-learn
numpy
jieba
uvicorn
gunicorn
starlette
fastapi
httpx
rank_bm25
bm25s[core]==0.2.0
rerankers==0.5.3
rouge-chinese
faiss-cpu
huggingface_hub
jionlp
loguru
flashtext
kafka-python 
python-snappy
sqlalchemy
zhon
pymysql
apscheduler
pytz
tzdata
streamlit
gradio>=5.0.0
sentence-transformers==3.0.1
accelerate
torch
torchvision
torchaudio
